Update Doubly_ended_Queue_using_Doubly_Linked_List
We can provide a more robust data structure if we do not permit users to
directly access or manipulate the nodes. In that way, we ensure that users
cannot invalidate the consistency of a list by mismanaging the linking of
nodes. A more subtle problem arises if a user were allowed to call the
insert between or delete node method of our DoublyLinkedBase class,
sending a node that does not belong to the given list as a parameter. (Go back
and look at that code and see why it causes a problem!)